refactor: create automation sieve after creating settings and switch to class methods for apis
